Import the Joy source code
The files included in this commit are the Joy source files as I found
them. However, they were all originally in the top level directory. I
blindly moved them into the lib/ and src/ directories, for the Joy and
C code, respectively, so I could start off with an organized file
structure. I have also removed a lot of other files, including
documentation and the BDW gc source code.

This will not build. My intention is to rewrite the documenation, and
include parts of the original documentation, specfically Thun's
writeups. I also plan to have the project to build by linking to the
gc library, rather than compiling a local copy.